+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### Version 0.6.1
 * Added a bug fix that prevents service tokens from auto-expiring.
+* Added a `ps` command to the `plugins` CLI to display information about the running plugins.
 
 #### Version 0.6.0 (March 5, 2017)
 * Added `calculate` interface to the `java` driver to perform aggregations.